requirements.txt
absl-py==2.4.0
colorama==0.4.6
filelock==3.25.2
fsspec==2026.3.0
grpcio==1.80.0
Jinja2==3.1.6
Markdown==3.10.2
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.4
opencv-python==4.13.0.92
packaging==26.0
pillow==12.2.0
protobuf==7.34.1
setuptools==81.0.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
typing_extensions==4.15.0
Werkzeug==3.1.8
